# src/main/resources/api/swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: ITAOBAO
basePath: /iTaobao
host: www.ym999.wang:8080
schemes:
  - http
  - https
paths:
...
  /history:
    get:
      tags: 
        - history
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有用户浏览历史记录
        
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - history
      # Describe this verb here. Note: you can use markdown
      description: |
        添加浏览历史记录数据
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/history'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - history
      description: |
        修改浏览历史记录信息
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/history'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - history
      description: |
        删除浏览历史记录数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: historyID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'    
  /man:
    get:
      tags: 
        - manager
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有管理员
        
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - manager
      # Describe this verb here. Note: you can use markdown
      description: |
        添加管理员数据
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/manager'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - manager
      description: |
        修改管理员信息
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/manager'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - manager
      description: |
        删除管理员数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: userID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'    
  /order:
    get:
      tags: 
        - order
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有订单
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - order
      # Describe this verb here. Note: you can use markdown
      description: |
        添加订单数据
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/orderForm'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - order
      description: |
        修改订单信息
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/orderForm'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - order
      description: |
        删除订单数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: orderID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'       
  /permission:
    get:
      tags: 
        - permission
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有权限
        
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - permission
      # Describe this verb here. Note: you can use markdown
      description: |
        添加权限数据
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/permission'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - permission
      description: |
        修改权限信息
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/permission'
      responses:
        # Response code
        200:
          description: Successful response
          # A schema describing your response object.
          # Use JSON Schema format
          schema:
            title: ArrayOfPersons
            type: array
            items:
              title: Person
              type: object
              properties:
                name:
                  type: string
                single:
                  type: boolean
    delete:
      tags:
        - permission
      description: |
        删除权限数据，单条数据，由传入的ID决定
      parameters:
        - 
          name: permissionID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'       
  /product:
    get:
      tags: 
        - product
      # Describe this verb here. Note: you can use markdown
      description: |
        查询所有产品
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/product'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - product
      # Describe this verb here. Note: you can use markdown
      description: |
        添加产品数据
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/product'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - product
      description: |
        修改产品信息
        
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/product'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - product
      description: |
        删除产品数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: userID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'       
  /reper:
    get:
      tags: 
        - reper
      # Describe this verb here. Note: you can use markdown
      description: |
         查询所有用户库存信息
         
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    post:
      tags:
        - reper
      # Describe this verb here. Note: you can use markdown
      description: |
        添加库存数据
      # This is array of GET operation parameters:
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/repertory'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    put:
      tags:
        - reper
      description: |
        修改库存信息
      parameters:
        -
          name: body
          in: body
          description: bus
          required: true
          schema:
            $ref: '#/definitions/repertory'
      responses:
        # Response code
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'
    delete:
      tags:
        - reper
      description: |
        删除库存数据，单条数据，由传入的ID决定
        
      parameters:
        - 
          name: repeID
          in: query
          description: name of user
          required: true
          type: integer
          format: int32
      responses:
  
        200:
          description: Successful response
          schema:
            $ref: '#/definitions/ApiResponse'       
...
definitions:
...
  history:
    type: object
    properties:
      historyID:
        type: integer
        format: int32
        description: l
      historyData:
        type: string
        description: ls
  manager:
    type: object
    properties:
      man_ID:
        type: integer
        format: int32
        description: l
      man_username:
        type: string
        description: l
      man_name:
        type: string
        description: the true name of this man
      man_password:
        type: string
        description: l
  orderForm:
    type: object
    properties:
      orderID:
        type: integer
        format: int32
        description: number of this order
      order_detail_ID:
        type: integer
        format: int32
        description: the id of this order's detail
      money:
        type: number
        format: float
        description: sl
      createTime:
        type: string
        format: date
        description: l
      payTime:
        type: string
        format: date
        description: pay time
      orderStatus:
        type: integer
        format: int32
        description: l
  permission:
    type: object
    properties:
      permissionID:
        type: integer
        format: int32
        description: id of permission
      permissionName:
        type: string
        description: name of permission
      permissionUrl:
        type: string
        description: l
  product:
    type: object
    properties:
      pro_ID:
        type: integer
        format: int32
        description: id
      pro_name:
        type: string
        description: name
      pro_img:
        type: string
        description: img
      pro_price:
        type: number
        format: float
        description: l
      pro_Info:
        type: string
        description: l
      pro_explain:
        type: string
        description: l
      pro_status:
        type: integer
        format: int32
        description: ls
  repertory:
    type: object
    properties:
      repe_ID:
        type: integer
        format: int32
        description: lll
      pro_ID:
        type: integer
        format: int32
        description: l
      repe_num:
        type: integer
        format: int32
        description: l
...
  ApiResponse:
    type: object
    properties:
      falg:
        type: boolean
        description: is sucessful?
      code:
        type: integer
        format: int32
      message:
        type: string
        description: return message